+++ ofrak_patch_maker/tests/test_x86_toolchain.py
@@ -11,7 +11,6 @@ from .toolchain_under_test import ToolchainUnderTest
 from .toolchain_c import (
     run_hello_world_test,
     run_bounds_check_test,
-    run_make_bom_parallel_test,
     run_relocatable_test,
 )
 from ofrak_type.architecture import (
@@ -65,7 +64,3 @@ def test_hello_world(toolchain_under_test: ToolchainUnderTest):
 
 def test_relocatable(toolchain_under_test: ToolchainUnderTest, tmp_path):
     run_relocatable_test(toolchain_under_test, tmp_path)
-
-
-def test_make_bom_parallel(toolchain_under_test: ToolchainUnderTest):
-    run_make_bom_parallel_test(toolchain_under_test)