+++ ofrak_patch_maker/CHANGELOG.md
@@ -5,8 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased](https://github.com/redballoonsecurity/ofrak/tree/master)
 
-### Added
-- Add a method for parallel generation of a BOM ([#743](https://github.com/redballoonsecurity/ofrak/pull/743))
+### Changed
+- make `make_bom` parallel using threads ([#743](https://github.com/redballoonsecurity/ofrak/pull/743))
 
 ## [4.1.0](https://github.com/redballoonsecurity/ofrak/compare/ofrak-patch-maker-v.4.0.2...ofrak-patch-maker-v.4.1.0) - 2025-10-03